🎈 perf(roller): 修正一个键鼠模式下可能存在的类型转换问题
activeRollerValue类型转换中防止错误转换为int32

+++ src/lib/hidio.c
@@ -208,7 +208,7 @@ void HIDIO_KBD_FreshData()
 {
   static uint16_t prevSpMseX = 0;
   // Roller
-  spMseData->x = ((~activeRollerValue) >> 1);
+  spMseData->x =  (uint16_t)(~activeRollerValue) >> 1;
 
   if (spMseData->x != prevSpMseX) {
     spMseChanged = 1;
